# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overflow Outcast - Virtual Reality Art Exhibition Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  email: string;
  username: string;
  createdAt: string;
}

export interface Profile {
  id: number;
  userId: number;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface ProfileInput {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface Gallery {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  roomTheme: string;
  published: boolean;
  slug: string;
  artworkCount: number;
  createdAt: string;
  roomSeed: number;
  roomMode: string;
  roomSize: number;
  decorationLevel: number;
  roomHeight: number;
  lightingMood: number;
}

export interface GalleryInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  roomTheme?: string;
  roomMode?: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  roomSize?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  decorationLevel?: number;
  roomSeed?: number;
}

export interface GalleryUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  roomTheme?: string;
  roomMode?: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  roomSize?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  decorationLevel?: number;
}

export interface GalleryRoomPatch {
  roomTheme?: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  roomSize?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  roomHeight?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  decorationLevel?: number;
  /**
     * @minimum 0.5
     * @maximum 2
     */
  lightingMood?: number;
  roomMode?: string;
}

export interface PublishToggle {
  published: boolean;
}

export interface GalleryDecoration {
  id: number;
  galleryId: number;
  type: string;
  x: number;
  z: number;
  rotY: number;
  createdAt: string;
}

export interface GalleryDecorationInput {
  type: string;
  x: number;
  z: number;
  rotY: number;
}

export interface GalleryDecorationPatch {
  x?: number;
  z?: number;
  rotY?: number;
}

export interface Artwork {
  id: number;
  galleryId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  /** @nullable */
  artistName?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  medium?: string | null;
  /** @nullable */
  dimensions?: string | null;
  xPosition: number;
  yPosition: number;
  zPosition: number;
  rotation: number;
  scale: number;
  isManuallyPlaced: boolean;
  createdAt: string;
}

export interface ArtworkInput {
  galleryId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  imageUrl: string;
  artistName?: string;
  year?: string;
  medium?: string;
  dimensions?: string;
  xPosition?: number;
  yPosition?: number;
  zPosition?: number;
  rotation?: number;
  scale?: number;
  isManuallyPlaced?: boolean;
}

export interface ArtworkUpdate {
  title?: string;
  description?: string;
  imageUrl?: string;
  artistName?: string;
  year?: string;
  medium?: string;
  dimensions?: string;
  xPosition?: number;
  yPosition?: number;
  zPosition?: number;
  rotation?: number;
  scale?: number;
  isManuallyPlaced?: boolean;
}

export interface ArtworkPlacementPatch {
  xPosition?: number;
  yPosition?: number;
  zPosition?: number;
  rotation?: number;
  scale?: number;
  isManuallyPlaced?: boolean;
}

export interface UploadResponse {
  url: string;
}

export interface PublicGallery {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  slug: string;
  roomTheme: string;
  roomSize: number;
  artworkCount: number;
  /** @nullable */
  artistName?: string | null;
  createdAt: string;
}

export interface PublicGalleryDetail {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  slug: string;
  roomTheme: string;
  roomSeed: number;
  roomMode: string;
  roomSize: number;
  roomHeight: number;
  decorationLevel: number;
  lightingMood: number;
  /** @nullable */
  artistName?: string | null;
  artworks: Artwork[];
}

export interface DashboardStats {
  totalGalleries: number;
  publishedGalleries: number;
  totalArtworks: number;
  recentGalleries: Gallery[];
}