# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Overflow Outcast - Virtual Reality Art Exhibition Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /galleries/{id}/decorations:
    get:
      operationId: listGalleryDecorations
      tags: [galleries]
      summary: List manual decorations for a gallery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of decorations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryDecoration"
    post:
      operationId: createGalleryDecoration
      tags: [galleries]
      summary: Place a decoration in the gallery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryDecorationInput"
      responses:
        "201":
          description: Created decoration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryDecoration"
        "404":
          description: Gallery not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /galleries/{id}/decorations/{decorId}:
    patch:
      operationId: updateGalleryDecoration
      tags: [galleries]
      summary: Update a placed decoration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: decorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryDecorationPatch"
      responses:
        "200":
          description: Updated decoration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryDecoration"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGalleryDecoration
      tags: [galleries]
      summary: Remove a placed decoration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: decorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /galleries/{id}/decorations/reset:
    delete:
      operationId: resetGalleryDecorations
      tags: [galleries]
      summary: Remove all manual decorations (restore seeded-RNG layout)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: All decorations removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /artworks/gallery/{galleryId}:
    get:
      operationId: listArtworks
      tags: [artworks]
      summary: List artworks in a gallery
      parameters:
        - name: galleryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of artworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artwork"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /artworks:
    post:
      operationId: createArtwork
      tags: [artworks]
      summary: Create an artwork in a gallery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkInput"
      responses:
        "201":
          description: Created artwork
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /artworks/{id}:
    put:
      operationId: updateArtwork
      tags: [artworks]
      summary: Update an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkUpdate"
      responses:
        "200":
          description: Updated artwork
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteArtwork
      tags: [artworks]
      summary: Delete an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /artworks/{id}/placement:
    patch:
      operationId: patchArtworkPlacement
      tags: [artworks]
      summary: Update artwork placement in room (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkPlacementPatch"
      responses:
        "200":
          description: Updated artwork
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GalleryDecoration:
      type: object
      properties:
        id:
          type: integer
        galleryId:
          type: integer
        type:
          type: string
        x:
          type: number
        z:
          type: number
        rotY:
          type: number
        createdAt:
          type: string
      required:
        - id
        - galleryId
        - type
        - x
        - z
        - rotY
        - createdAt
    GalleryDecorationInput:
      type: object
      properties:
        type:
          type: string
        x:
          type: number
        z:
          type: number
        rotY:
          type: number
      required:
        - type
        - x
        - z
        - rotY
    GalleryDecorationPatch:
      type: object
      properties:
        x:
          type: number
        z:
          type: number
        rotY:
          type: number
    Artwork:
      type: object
      properties:
        id:
          type: integer
        galleryId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: string
        artistName:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        medium:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        xPosition:
          type: number
        yPosition:
          type: number
        zPosition:
          type: number
        rotation:
          type: number
        scale:
          type: number
        isManuallyPlaced:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - galleryId
        - title
        - imageUrl
        - xPosition
        - yPosition
        - zPosition
        - rotation
        - scale
        - isManuallyPlaced
        - createdAt
    ArtworkInput:
      type: object
      properties:
        galleryId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: string
        artistName:
          type: string
        year:
          type: string
        medium:
          type: string
        dimensions:
          type: string
        xPosition:
          type: number
        yPosition:
          type: number
        zPosition:
          type: number
        rotation:
          type: number
        scale:
          type: number
        isManuallyPlaced:
          type: boolean
      required:
        - galleryId
        - title
        - imageUrl
    ArtworkUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        artistName:
          type: string
        year:
          type: string
        medium:
          type: string
        dimensions:
          type: string
        xPosition:
          type: number
        yPosition:
          type: number
        zPosition:
          type: number
        rotation:
          type: number
        scale:
          type: number
        isManuallyPlaced:
          type: boolean
    ArtworkPlacementPatch:
      type: object
      properties:
        xPosition:
          type: number
        yPosition:
          type: number
        zPosition:
          type: number
        rotation:
          type: number
        scale:
          type: number
        isManuallyPlaced:
          type: boolean
...