# lib/api-zod/src/generated/types/artworkPlacementPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overflow Outcast - Virtual Reality Art Exhibition Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ArtworkPlacementPatch {
  xPosition?: number;
  yPosition?: number;
  zPosition?: number;
  rotation?: number;
  scale?: number;
  isManuallyPlaced?: boolean;
}