# lib/api-zod/src/generated/types/galleryDecorationPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overflow Outcast - Virtual Reality Art Exhibition Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryDecorationPatch {
  x?: number;
  z?: number;
  rotY?: number;
}